+++ example/java/example/src/test/java/io/bitmax/api/examples/account/rest/GetAccountOrdersExample.java
@@ -1,8 +1,8 @@
-package io.bitmax.api.examples.account;
+package io.bitmax.api.examples.account.rest;
 
 import io.bitmax.api.rest.client.BitMaxRestApiClientAccount;
 
-public class AccountGroupRest {
+public class GetAccountOrdersExample {
     public static void main(String[] args) {
         String apiKey = "<apikey>";
         String secret = "<secret>";
@@ -10,6 +10,6 @@ public class AccountGroupRest {
 
         BitMaxRestApiClientAccount restClient = new BitMaxRestApiClientAccount(apiKey, secret, baseUrl);
 
-        System.out.println(restClient.getUserInfo());
+        System.out.println(restClient.getOpenOrders());
     }
 }
