+++ example/java/example/src/test/java/io/bitmax/api/examples/common/rest/BarHistExample.java
@@ -1,4 +1,4 @@
-package io.bitmax.api.examples.common;
+package io.bitmax.api.examples.common.rest;
 
 import io.bitmax.api.rest.client.BitMaxRestApiClient;
 import io.bitmax.api.rest.client.Interval;
@@ -6,7 +6,7 @@ import io.bitmax.api.rest.messages.responses.BarHist;
 
 import java.util.Arrays;
 
-public class BarHistRestExample {
+public class BarHistExample {
     public static void main(String[] args) {
 
         BitMaxRestApiClient restClient = new BitMaxRestApiClient();
